move fragmenturi outside utils
