+++ lib/fragment.js
@@ -5,7 +5,6 @@ const PassThrough = require('stream').PassThrough;
 const zlib = require('zlib');
 const ContentLengthStream = require('./streams/content-length-stream');
 const parseLinkHeader = require('./parse-link-header');
-const { getFragmentAssetUris } = require('./utils');
 
 const { globalTracer, Tags } = require('opentracing');
 const tracer = globalTracer();
@@ -15,6 +14,20 @@ const hasValue = value => {
         return true;
     }
     return false;
+};
+
+const getFragmentAssetUris = (refs, assetSize) => {
+    const scriptUris = [];
+    const styleUris = [];
+
+    for (const ref of refs) {
+        if (ref.rel === 'fragment-script') {
+            scriptUris.push(ref.uri);
+        } else if (ref.rel === 'stylesheet') {
+            styleUris.push(ref.uri);
+        }
+    }
+    return [scriptUris.slice(0, assetSize), styleUris.slice(0, assetSize)];
 };
 
 /**
