+++ lib/utils.js
@@ -68,20 +68,6 @@ const getFragmentAssetsToPreload = (styleRefs, scriptRefs, { host } = {}) => {
     return assetsToPreload;
 };
 
-function getFragmentAssetUris(refs, assetSize) {
-    const scriptUris = [];
-    const styleUris = [];
-
-    for (const ref of refs) {
-        if (ref.rel === 'fragment-script') {
-            scriptUris.push(ref.uri);
-        } else if (ref.rel === 'stylesheet') {
-            styleUris.push(ref.uri);
-        }
-    }
-    return [scriptUris.slice(0, assetSize), styleUris.slice(0, assetSize)];
-}
-
 const nextIndexGenerator = (initialIndex, step) => {
     let index = initialIndex;
 
@@ -96,6 +82,5 @@ module.exports = {
     getCrossOrigin,
     getFragmentAssetsToPreload,
     getLoaderScript,
-    getFragmentAssetUris,
     nextIndexGenerator
 };
